Revert "Parallelize data fetching tasks in the driver (#3450)"
This reverts commit 85e0d3a321d348df8ef851d860d596717e9b41da.
