+++ crates/driver/src/domain/competition/mod.rs
@@ -5,21 +5,17 @@ use {
         time::{self, Remaining},
     },
     crate::{
-        domain::{
-            competition::{solution::Settlement, sorting::SortingStrategy},
-            eth,
-            time::DeadlineExceeded,
-        },
+        domain::{competition::solution::Settlement, eth, time::DeadlineExceeded},
         infra::{
             self,
             Simulator,
             blockchain::Ethereum,
             notify,
-            observe::{self, metrics},
+            observe,
             simulator::{RevertError, SimulatorError},
             solver::{self, SolutionMerging, Solver},
         },
-        util::{Bytes, math},
+        util::Bytes,
     },
     futures::{StreamExt, future::Either, stream::FuturesUnordered},
     itertools::Itertools,
@@ -31,26 +27,23 @@ use {
         time::Duration,
     },
     tap::TapFallible,
-    tokio::{
-        sync::{mpsc, oneshot},
-        task,
-    },
+    tokio::sync::{mpsc, oneshot},
     tracing::{Instrument, instrument},
 };
 
 pub mod auction;
 pub mod bad_tokens;
 pub mod order;
-mod pre_processing;
 pub mod solution;
-pub mod sorting;
+mod sorting;
 
-pub use {auction::Auction, order::Order, pre_processing::DataAggregator, solution::Solution};
+pub use {
+    auction::{Auction, AuctionProcessor},
+    order::Order,
+    solution::Solution,
+};
 
 use crate::domain::BlockNo;
-
-type BalanceGroup = (order::Trader, eth::TokenAddress, order::SellTokenBalance);
-type Balances = HashMap<BalanceGroup, order::SellAmount>;
 
 /// An ongoing competition. There is one competition going on per solver at any
 /// time. The competition stores settlements to solutions generated by the
@@ -67,13 +60,10 @@ pub struct Competition {
     /// Cached solutions with the most recent solutions at the front.
     pub settlements: Mutex<VecDeque<Settlement>>,
     pub bad_tokens: Arc<bad_tokens::Detector>,
-    fetcher: Arc<pre_processing::DataAggregator>,
     settle_queue: mpsc::Sender<SettleRequest>,
-    order_sorting_strategies: Vec<Arc<dyn sorting::SortingStrategy>>,
 }
 
 impl Competition {
-    #[allow(clippy::too_many_arguments)]
     pub fn new(
         solver: Solver,
         eth: Ethereum,
@@ -81,8 +71,6 @@ impl Competition {
         simulator: Simulator,
         mempools: Mempools,
         bad_tokens: Arc<bad_tokens::Detector>,
-        fetcher: Arc<DataAggregator>,
-        order_sorting_strategies: Vec<Arc<dyn sorting::SortingStrategy>>,
     ) -> Arc<Self> {
         let (settle_sender, settle_receiver) = mpsc::channel(solver.settle_queue_size());
 
@@ -95,8 +83,6 @@ impl Competition {
             settlements: Default::default(),
             settle_queue: settle_sender,
             bad_tokens,
-            fetcher,
-            order_sorting_strategies,
         });
 
         let competition_clone = Arc::clone(&competition);
@@ -112,58 +98,15 @@ impl Competition {
     /// Solve an auction as part of this competition.
     #[instrument(skip_all)]
     pub async fn solve(&self, auction: Auction) -> Result<Option<Solved>, Error> {
-        let auction = Arc::new(auction);
-        let tasks = self.fetcher.start_or_get_tasks_for_auction(auction.clone());
-        let mut auction = Arc::unwrap_or_clone(auction);
-
-        let settlement_contract = self.eth.contracts().settlement().address();
-        let solver_address = self.solver.account().address();
-        let order_sorting_strategies = self.order_sorting_strategies.clone();
-
-        // Add the CoW AMM orders to the auction
-        let cow_amm_orders = tasks.cow_amm_orders.await;
-        auction.orders.extend(cow_amm_orders.iter().cloned());
-
-        let sort_orders_future = Self::run_blocking_with_timer("sort_orders", move || {
-            // Use spawn_blocking() because a lot of CPU bound computations are happening
-            // and we don't want to block the runtime for too long.
-            Self::sort_orders(
-                auction,
-                solver_address,
-                order_sorting_strategies,
-                settlement_contract,
-            )
-        });
-
-        // We can sort the orders and fetch auction data in parallel
-        let (auction, balances, app_data) =
-            tokio::join!(sort_orders_future, tasks.balances, tasks.app_data);
-
-        let auction = Self::run_blocking_with_timer("update_orders", move || {
-            // Same as before with sort_orders, we use spawn_blocking() because a lot of CPU
-            // bound computations are happening and we want to avoid blocking
-            // the runtime.
-            Self::update_orders(
-                auction,
-                balances,
-                app_data,
-                cow_amm_orders,
-                &eth::Address(settlement_contract),
-            )
-        })
-        .await;
-
-        // We can tun bad token filtering and liquidity fetching in parallel
+        let pairs_to_fetch = match self.solver.liquidity() {
+            solver::Liquidity::Fetch => auction.liquidity_pairs(),
+            solver::Liquidity::Skip => Default::default(),
+        };
         let (liquidity, auction) = tokio::join!(
-            async {
-                match self.solver.liquidity() {
-                    solver::Liquidity::Fetch => tasks.liquidity.await,
-                    solver::Liquidity::Skip => Arc::new(Vec::new()),
-                }
-            },
+            self.liquidity
+                .fetch(&pairs_to_fetch, self.solver.fetch_liquidity_at_block()),
             self.without_unsupported_orders(auction)
         );
-
         let auction = &auction;
 
         // Fetch the solutions from the solver.
@@ -369,149 +312,6 @@ impl Competition {
         }
 
         Ok(score)
-    }
-
-    // Oders already need to be sorted from most relevant to least relevant so that
-    // we allocate balances for the most relevants first.
-    fn sort_orders(
-        mut auction: Auction,
-        solver: eth::H160,
-        order_sorting_strategies: Vec<Arc<dyn SortingStrategy>>,
-        _settlement_contract: eth::H160,
-    ) -> Auction {
-        sorting::sort_orders(
-            &mut auction.orders,
-            &auction.tokens,
-            &solver,
-            &order_sorting_strategies,
-        );
-        auction
-    }
-
-    /// Removes orders that cannot be filled due to missing funds of the owner
-    /// and updates the fetched app data.
-    /// It allocates available funds from left to right so the orders should
-    /// already be sorted by priority going in.
-    fn update_orders(
-        mut auction: Auction,
-        balances: Arc<Balances>,
-        app_data: Arc<HashMap<order::app_data::AppDataHash, app_data::ValidatedAppData>>,
-        cow_amm_orders: Arc<Vec<Order>>,
-        settlement_contract: &eth::Address,
-    ) -> Auction {
-        // Clone balances since we only aggregate data once but each solver needs
-        // to use and modify the data individually.
-        let mut balances = balances.as_ref().clone();
-        let cow_amms: HashSet<_> = cow_amm_orders.iter().map(|o| o.uid).collect();
-
-        // The auction that we receive from the `autopilot` assumes that there
-        // is sufficient balance to completely cover all the orders. **This is
-        // not the case** (as the protocol should not chose which limit orders
-        // get filled for some given sell token balance). This loop goes through
-        // the priority sorted orders and allocates the available user balance
-        // to each order, and potentially scaling the order's `available` amount
-        // down in case the available user balance is only enough to partially
-        // cover the rest of the order.
-        auction.orders.retain_mut(|order| {
-            if cow_amms.contains(&order.uid) {
-                // cow amm orders already get constructed fully initialized
-                // so we don't have to handle them here anymore.
-                // Without this short circuiting logic they would get filtered
-                // out later because we don't bother fetching their balances
-                // for performance reasons.
-                return true;
-            }
-
-            // Update order app data if it was fetched.
-            if let Some(fetched_app_data) = app_data.get(&order.app_data.hash()) {
-                order.app_data = fetched_app_data.clone().into();
-                if order.app_data.flashloan().is_some() {
-                    // If an order requires a flashloan we assume all the necessary
-                    // sell tokens will come from there. But the receiver must be the
-                    // settlement contract because that is how the driver expects
-                    // the flashloan to be repaid for now.
-                    return order.receiver.as_ref() == Some(settlement_contract);
-                }
-            }
-
-            let remaining_balance = match balances.get_mut(&(
-                order.trader(),
-                order.sell.token,
-                order.sell_token_balance,
-            )) {
-                Some(balance) => balance,
-                None => {
-                    let reason = observe::OrderExcludedFromAuctionReason::CouldNotFetchBalance;
-                    observe::order_excluded_from_auction(order, reason);
-                    return false;
-                }
-            };
-
-            let max_sell = order::SellAmount(order.available().sell.amount.0);
-
-            let allocated_balance = match order.partial {
-                order::Partial::Yes { .. } => max_sell.min(*remaining_balance),
-                order::Partial::No if max_sell <= *remaining_balance => max_sell,
-                _ => order::SellAmount::default(),
-            };
-            if allocated_balance.0.is_zero() {
-                observe::order_excluded_from_auction(
-                    order,
-                    observe::OrderExcludedFromAuctionReason::InsufficientBalance,
-                );
-                return false;
-            }
-
-            // We need to scale the available amount in the order based on
-            // allocated balance. We cannot naively just set the `available`
-            // amount to equal the `allocated_balance` because of two reasons:
-            // 1. They are in different units. `available` is a `TargetAmount` which means
-            //    it would be in buy token for buy orders and not in sell token like the
-            //    `allocated_balance`
-            // 2. Account for fees. Even in the case of sell orders, `available` is
-            //    potentially different to `allocated_balance` because of fee scaling. For
-            //    example, imagine a partially fillable order selling 100 tokens with a fee
-            //    of 10 for a user with a balance of 50. The `allocated_balance` would be 50
-            //    tokens, but the `available` amount needs to be less! We want the
-            //    following: `available + (fee * available / sell) <= allocated_balance`
-            if let order::Partial::Yes { available } = &mut order.partial {
-                *available = order::TargetAmount(
-                    math::mul_ratio(available.0, allocated_balance.0, max_sell.0)
-                        .unwrap_or_default(),
-                );
-            }
-            if order.available().is_zero() {
-                observe::order_excluded_from_auction(
-                    order,
-                    observe::OrderExcludedFromAuctionReason::OrderWithZeroAmountRemaining,
-                );
-                return false;
-            }
-
-            remaining_balance.0 -= allocated_balance.0;
-
-            true
-        });
-
-        auction
-    }
-
-    /// Runs a blocking function on a background thread, timing it using the
-    /// given stage name.
-    pub async fn run_blocking_with_timer<T, F>(stage: &'static str, f: F) -> T
-    where
-        F: FnOnce() -> T + Send + 'static,
-        T: Send + 'static,
-    {
-        task::spawn_blocking(move || {
-            let _timer = metrics::get().processing_stage_timer(stage);
-            f()
-        })
-        .await
-        .expect(
-            "Either runtime was shut down before spawning the task or no OS threads are \
-             available; no sense in handling those errors",
-        )
     }
 
     pub async fn reveal(
